gcp_weather_pulse/dashboard/requirements.txt
streamlit==1.37.1
google-cloud-bigquery[pandas]==3.25.0
google-cloud-bigquery-storage>=2.25.0
db-dtypes>=1.3.0
pandas==2.2.2
pyarrow==17.0.0